Refactor transaction and gold models to use Kotlin types
- Updated `price` and `amount` fields in `DomainGold`, `DomainTransaction`, and `CreateTransactionReq` to use `Int` and `Double` instead of `BigDecimal`.
- Modified related UI components to reflect changes in data types for price and amount.
- Introduced a new `TransactionListItem` composable for displaying transaction details.
- Enhanced `UpsertTransactionScreen` to handle date selection with a custom date picker dialog.
- Updated preview functions to align with new data types.
- Added error handling for unexpected client errors with a user-friendly message.
- Adjusted Gradle properties for improved performance.
- Updated OpenAPI documentation to reflect changes in request and response models.

+++ android/app/src/main/java/id/my/rizalanggoro/arta/feature/gold/presentation/detail/GoldDetailScreen.kt
@@ -175,7 +175,7 @@ private fun GoldDetailScreenPreview() {
                     walletId = 2,
                     date = "2026-05-16",
                     grams = BigDecimal.valueOf(10.0),
-                    price = BigDecimal.valueOf(9000000.0),
+                    price = 9000000,
                     type = "pure_gold",
                     carat = BigDecimal.valueOf(24.0),
                     notes = "Contoh catatan",
